Moved unit tests into own file. #18
